+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added 3 tutorials and embended existing colab tutorial in it.
 - Support artifacts for each optimizer based on evaluated trials.
 - Add a generic structure for the filesystem writers.
+- Add live_plots flag for run() API to print the generic and optimizer-specific plots
 
 ### Fixed
 - fix device mismatch in the tensors created in gp acqisition process